Removed methods
Methods from sam_template_validator were removed. tests were moved and fixed accordingly. Load policies is now done in wrapper class.
