+++ samcli/commands/validate/lib/sam_template_validator.py
@@ -9,7 +9,7 @@ from samtranslator.parser import parser
 from samtranslator.translator.translator import Translator
 from boto3.session import Session
 
-from samcli.lib.replace_uri.replace_uri import _replace_local_codeuri as external_replace_local_codeuri
+from samcli.lib.replace_uri.replace_uri import replace_local_codeuri as external_replace_local_codeuri
 from samcli.lib.replace_uri.replace_uri import is_s3_uri as external_is_s3_uri
 
 from samcli.yamlhelper import yaml_dump
@@ -55,7 +55,7 @@ class SamTemplateValidator:
         InvalidSamDocumentException
              If the template is not valid, an InvalidSamDocumentException is raised
         """
-        managed_policy_map = self.managed_policy_loader.load()
+        managed_policy_map = self.managed_policy_loader
 
         sam_translator = Translator(
             managed_policy_map=managed_policy_map,
@@ -64,7 +64,7 @@ class SamTemplateValidator:
             boto_session=self.boto3_session,
         )
 
-        self.sam_template = self._replace_local_codeuri()
+        self.sam_template = external_replace_local_codeuri(self.sam_template)
 
         try:
             template = sam_translator.translate(sam_template=self.sam_template, parameter_values={})
@@ -73,27 +73,6 @@ class SamTemplateValidator:
             raise InvalidSamDocumentException(
                 functools.reduce(lambda message, error: message + " " + str(error), e.causes, str(e))
             ) from e
-
-    def _replace_local_codeuri(self):
-        return external_replace_local_codeuri(self.sam_template)
-
-    @staticmethod
-    def is_s3_uri(uri):
-        """
-        Checks the uri and determines if it is a valid S3 Uri
-
-        Parameters
-        ----------
-        uri str, required
-            Uri to check
-
-        Returns
-        -------
-        bool
-            Returns True if the uri given is an S3 uri, otherwise False
-
-        """
-        return external_is_s3_uri(uri)
 
     @staticmethod
     def _update_to_s3_uri(property_key, resource_property_dict, s3_uri_value="s3://bucket/value"):
