+++ tests/unit/commands/validate/lib/test_sam_template_validator.py
@@ -1,7 +1,6 @@
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
-from samcli.lib.utils.packagetype import IMAGE
 from samtranslator.public.exceptions import InvalidDocumentException
 
 from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
@@ -13,8 +12,7 @@ class TestSamTemplateValidator(TestCase):
     @patch("samcli.commands.validate.lib.sam_template_validator.Translator")
     @patch("samcli.commands.validate.lib.sam_template_validator.parser")
     def test_is_valid_returns_true(self, sam_parser, sam_translator, boto_session_patch):
-        managed_policy_mock = Mock()
-        managed_policy_mock.load.return_value = {"policy": "SomePolicy"}
+        managed_policy_mock = {"policy": "SomePolicy"}
         template = {"a": "b"}
 
         parser = Mock()
@@ -43,8 +41,7 @@ class TestSamTemplateValidator(TestCase):
     @patch("samcli.commands.validate.lib.sam_template_validator.Translator")
     @patch("samcli.commands.validate.lib.sam_template_validator.parser")
     def test_is_valid_raises_exception(self, sam_parser, sam_translator, boto_session_patch):
-        managed_policy_mock = Mock()
-        managed_policy_mock.load.return_value = {"policy": "SomePolicy"}
+        managed_policy_mock = {"policy": "SomePolicy"}
         template = {"a": "b"}
 
         parser = Mock()
@@ -82,15 +79,6 @@ class TestSamTemplateValidator(TestCase):
         # check to see if SamParser was created
         self.assertIsNotNone(validator.sam_parser)
 
-    def test_uri_is_s3_uri(self):
-        self.assertTrue(SamTemplateValidator.is_s3_uri("s3://bucket/key"))
-
-    def test_uri_is_not_s3_uri(self):
-        self.assertFalse(SamTemplateValidator.is_s3_uri("www.amazon.com"))
-
-    def test_int_is_not_s3_uri(self):
-        self.assertFalse(SamTemplateValidator.is_s3_uri(100))
-
     def test_update_to_s3_uri_with_non_s3_uri(self):
         property_value = {"CodeUri": "somevalue"}
         SamTemplateValidator._update_to_s3_uri("CodeUri", property_value)
@@ -110,199 +98,3 @@ class TestSamTemplateValidator(TestCase):
         SamTemplateValidator._update_to_s3_uri("CodeUri", property_value)
 
         self.assertEqual(property_value.get("CodeUri"), "s3://bucket/key/version")
-
-    def test_replace_local_codeuri(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Resources": {
-                "ServerlessApi": {
-                    "Type": "AWS::Serverless::Api",
-                    "Properties": {"StageName": "Prod", "DefinitionUri": "./"},
-                },
-                "ServerlessFunction": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"Handler": "index.handler", "CodeUri": "./", "Runtime": "nodejs6.10", "Timeout": 60},
-                },
-                "ServerlessLayerVersion": {"Type": "AWS::Serverless::LayerVersion", "Properties": {"ContentUri": "./"}},
-                "ServerlessStateMachine": {
-                    "Type": "AWS::Serverless::StateMachine",
-                    "Properties": {"DefinitionUri": "./", "Role": "test-role-arn"},
-                },
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        template_resources = validator.sam_template.get("Resources")
-        self.assertEqual(
-            template_resources.get("ServerlessApi").get("Properties").get("DefinitionUri"), "s3://bucket/value"
-        )
-        self.assertEqual(
-            template_resources.get("ServerlessFunction").get("Properties").get("CodeUri"), "s3://bucket/value"
-        )
-        self.assertEqual(
-            template_resources.get("ServerlessLayerVersion").get("Properties").get("ContentUri"), "s3://bucket/value"
-        )
-        self.assertEqual(
-            template_resources.get("ServerlessStateMachine").get("Properties").get("DefinitionUri"), "s3://bucket/value"
-        )
-
-    def test_replace_local_codeuri_when_no_codeuri_given(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Resources": {
-                "ServerlessApi": {"Type": "AWS::Serverless::Api", "Properties": {"StageName": "Prod"}},
-                "ServerlessFunction": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"Handler": "index.handler", "Runtime": "nodejs6.10", "Timeout": 60},
-                },
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        tempalte_resources = validator.sam_template.get("Resources")
-        self.assertEqual(
-            tempalte_resources.get("ServerlessFunction").get("Properties").get("CodeUri"), "s3://bucket/value"
-        )
-
-    def test_dont_replace_local_codeuri_when_no_codeuri_given_packagetype_image(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Resources": {
-                "ServerlessApi": {"Type": "AWS::Serverless::Api", "Properties": {"StageName": "Prod"}},
-                "ServerlessFunction": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"PackageType": IMAGE, "ImageUri": "myimage:latest", "Timeout": 60},
-                },
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        template_resources = validator.sam_template.get("Resources")
-        self.assertEqual(
-            template_resources.get("ServerlessFunction").get("Properties").get("CodeUri", "NotPresent"), "NotPresent"
-        )
-
-    def test_dont_replace_codeuri_when_global_code_uri_given_packagetype_image(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Globals": {"Function": {"CodeUri": "globalcodeuri", "Timeout": "3"}},
-            "Resources": {
-                "ServerlessApi": {"Type": "AWS::Serverless::Api", "Properties": {"StageName": "Prod"}},
-                "ServerlessFunction": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"PackageType": IMAGE, "ImageUri": "myimage:latest", "Timeout": 60},
-                },
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        template_resources = validator.sam_template.get("Resources")
-        self.assertEqual(
-            template_resources.get("ServerlessFunction").get("Properties").get("CodeUri", "NotPresent"), "NotPresent"
-        )
-
-    def test_dont_replace_codeuri_when_global_code_uri_given__both_packagetype(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Globals": {
-                "Function": {
-                    "CodeUri": "s3://globalcodeuri",
-                }
-            },
-            "Resources": {
-                "ServerlessApi": {"Type": "AWS::Serverless::Api", "Properties": {"StageName": "Prod"}},
-                "ServerlessFunctionImage": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"PackageType": IMAGE, "ImageUri": "myimage:latest", "Timeout": 60},
-                },
-                "ServerlessFunctionZip": {
-                    "Type": "AWS::Serverless::Function",
-                    "Properties": {"Handler": "index.handler", "Runtime": "nodejs6.10", "Timeout": 60},
-                },
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        template_resources = validator.sam_template.get("Resources")
-        self.assertEqual(
-            template_resources.get("ServerlessFunctionImage").get("Properties").get("CodeUri", "NotPresent"),
-            "NotPresent",
-        )
-        # Globals not set since they cant apply to both Zip and Image based packagetypes.
-        self.assertEqual(
-            template_resources.get("ServerlessFunctionZip").get("Properties").get("CodeUri"), "s3://bucket/value"
-        )
-
-    def test_DefinitionUri_does_not_get_added_to_template_when_DefinitionBody_given(self):
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Resources": {
-                "ServerlessApi": {
-                    "Type": "AWS::Serverless::Api",
-                    "Properties": {"StageName": "Prod", "DefinitionBody": {"swagger": {}}},
-                }
-            },
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        tempalte_resources = validator.sam_template.get("Resources")
-        self.assertNotIn("DefinitionUri", tempalte_resources.get("ServerlessApi").get("Properties"))
-        self.assertIn("DefinitionBody", tempalte_resources.get("ServerlessApi").get("Properties"))
-
-    def test_replace_local_codeuri_with_no_resources(self):
-
-        template = {
-            "AWSTemplateFormatVersion": "2010-09-09",
-            "Transform": "AWS::Serverless-2016-10-31",
-            "Resources": {},
-        }
-
-        managed_policy_mock = Mock()
-
-        validator = SamTemplateValidator(template, managed_policy_mock)
-
-        validator._replace_local_codeuri()
-
-        # check template
-        self.assertEqual(validator.sam_template.get("Resources"), {})
